refactor: nuke metrics.go
